+++ dev/App.vue
@@ -3,6 +3,11 @@
     <v-md-editor
       v-model="text"
       height="500px"
+      autofocus
+      :disabled-menus="[]"
+      @upload-image="handleUploadImage"
+      @fullscreen-change="handleFullscreenChange"
+      @save="handleSave"
       ref="editor"
     />
   </div>
@@ -14,16 +19,22 @@ import text from './text';
 export default {
   data () {
     return {
-      text: '',
+      text,
     };
   },
   methods: {
-    handleImageUpload(e, insertImage) {
-      console.log(e);
+    handleFullscreenChange (v) {
+      console.log(v);
+    },
+    handleUploadImage(e, insertImage, files) {
+      console.log(files);
 
       insertImage({
         desc: '111',
       });
+    },
+    handleSave (v, html) {
+      console.log(v, html);
     },
   },
 };