+++ docs/.vuepress/components/upload-image.vue
@@ -28,9 +28,9 @@ export default {
     }
   },
   methods: {
-    handleUploadImage (event, insertImage) {
-      const { files } = event.target;
-      // 拿到 File 对象之后上传到文件服务器，然后向编辑框中插入对应的内容
+    handleUploadImage (event, insertImage, files) {
+      // 拿到 files 之后上传到文件服务器，然后向编辑框中插入对应的内容
+      console.log(files);
       
       insertImage({
         url: 'https://ss0.bdstatic.com/70cFvHSh_Q1YnxGkpoWK1HF6hhy/it/u=1269952892,3525182336&fm=26&gp=0.jpg',