+++ src/base-editor.vue
@@ -8,7 +8,7 @@
     :fullscreen="fullscreen"
     :left-area-visible="tocVisible"
     left-area-title="目录导航"
-    :mode="mode"
+    :mode="currentMode"
     @editor-wrapper-click="handleEditorWrapperClick"
     @toolbar-item-click="handleToolbarItemClick"
     @toolbar-menu-click="handleToolbarMenuClick"
@@ -27,14 +27,15 @@
     >
       <v-md-textarea-editor
         :value="text"
+        :placeholder="placeholder"
         @input="handleInput"
         @click.native.stop
+        @drop.native="handleDrop"
         ref="editorEgine"
       />
     </scrollbar>
     <scrollbar
       slot="preview"
-      :disabled="isPreviewMode"
       ref="previewScroller"
     >
       <v-md-preview
@@ -44,10 +45,10 @@
         ref="preview"
       />
     </scrollbar>
-    <v-md-upload-img
+    <v-md-upload-file
       v-if="hasUploadImage"
-      :upload-image-config="$attrs['upload-image-config'] || $attrs.uploadImageConfig"
-      ref="uploadImage"
+      :upload-config="uploadConfig"
+      ref="uploadFile"
     />
   </v-md-container>
 </template>
@@ -61,11 +62,6 @@ const component = {
   components: {
     [TextareaEditor.name]: TextareaEditor,
   },
-  computed: {
-    editorEgine () {
-      return this.$refs.editorEgine;
-    },
-  },
   watch: {
     value() {
       this.text = this.value;
@@ -73,24 +69,29 @@ const component = {
   },
   methods: {
     handleEditorWrapperClick() {
+      this.setFocusEnd();
+    },
+    // Must implement
+    editorFocusEnd () {
       this.focus();
 
-      this.editorEgine.setRange({
+      this.$refs.editorEgine.setRange({
         start: this.text.length,
         end: this.text.length,
       });
     },
     // Must implement
     delLineLeft () {
-      const { start } = this.editorEgine.getRange();
+      const { editorEgine } = this.$refs;
+      const { start } = editorEgine.getRange();
 
       const leftText = this.getCursorLineLeftText();
-      this.editorEgine.setRange({ start: start - leftText.length - 1, end: start });
+      editorEgine.setRange({ start: start - leftText.length - 1, end: start });
       this.replaceSelectionText('\n');
     },
     // Must implement
     getCursorLineLeftText () {
-      const { start, end } = this.editorEgine.getRange();
+      const { start, end } = this.$refs.editorEgine.getRange();
 
       return start === end ? this.text.slice(0, start).split('\n').pop() : null;
     },
@@ -108,19 +109,19 @@ const component = {
     },
     // Must implement
     heightAtLine (...arg) {
-      return this.editorEgine.heightAtLine(...arg);
+      return this.$refs.editorEgine.heightAtLine(...arg);
     },
     // Must implement
     focus() {
-      this.editorEgine.focus();
+      this.$refs.editorEgine.focus();
     },
     // Must implement
     undo() {
-      this.editorEgine.undo();
+      this.$refs.editorEgine.undo();
     },
     // Must implement
     redo() {
-      this.editorEgine.redo();
+      this.$refs.editorEgine.redo();
     },
     // Must implement
     clear() {
@@ -130,11 +131,11 @@ const component = {
     },
     // Must implement
     replaceSelectionText(text) {
-      this.editorEgine.insertText(text);
+      this.$refs.editorEgine.insertText(text);
     },
     // Must implement
     getCurrentSelectedStr () {
-      const { start, end } = this.editorEgine.getRange();
+      const { start, end } = this.$refs.editorEgine.getRange();
 
       return end > start ? this.text.slice(start, end) : null;
     },
@@ -151,7 +152,7 @@ const component = {
       const rangeStartIndex = insertTextIndex + selectedIndexOfStr;
       const rangeEndIndex = rangeStartIndex + selectedText.length;
 
-      this.editorEgine.setRange({
+      this.$refs.editorEgine.setRange({
         start: rangeStartIndex,
         end: rangeEndIndex,
       });