+++ src/toolbar/save.js
@@ -1,7 +1,7 @@
 export default {
   name: 'save',
   icon: 'v-md-icon-save',
-  title: '保存',
+  title: '保存（Ctrl+S）',
   action(editor) {
     editor.save();
   },